intermediate change: merge normalization code into normalize_flow.go
